==> notebooks/1_data_collection.ipynb <==
import psycopg2
import sys, os
import numpy as np
import pandas as pd
# import example_psql as creds
import pandas.io.sql as psql

import credentials

sys.path.append(os.path.abspath('../helper/sql_helper.py'))
import sql_helper

tables = ["flights","passengers","fuel_comsumption","flights_test"]

con = psycopg2.connect(database=credentials.REMOTE["PGDATABASE"], user=credentials.REMOTE["PGUSER"], password=credentials.REMOTE["PGPASSWORD"], 
                       host=credentials.REMOTE["PGHOST"], port=credentials.REMOTE["PGPORT"])

if con!=None:
    row_count = sql_helper.get_number_of_rows(tables,con)
    
con.close()

row_count 

con = psycopg2.connect(database=credentials.REMOTE["PGDATABASE"], user=credentials.REMOTE["PGUSER"], password=credentials.REMOTE["PGPASSWORD"], 
                       host=credentials.REMOTE["PGHOST"], port=credentials.REMOTE["PGPORT"])

if con!=None:
    dataframes = sql_helper.get_df_from_sql(tables,con,number_of_rows=None)
    
con.close()

for table_name,data in zip(tables,dataframes):
    data.to_csv(f"{table_name}.csv")
