==> notebooks/2_generate_sample.ipynb <==
# Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Read file

flights_master = pd.read_csv(r"../data/flights_2019.csv")
flights_master.head()

flights_master.shape

# Sample size reduction

flights = flights_master.copy()

flights.loc[:,["fl_date","crs_arr_time","crs_dep_time"]] = flights[["fl_date","crs_arr_time","crs_dep_time"]].astype('string')

arr_time = pd.to_datetime(flights["fl_date"]+" "+ flights["crs_arr_time"].str.zfill(4).replace({'2400':'0000'}))
dep_time = pd.to_datetime(flights["fl_date"]+" "+ flights["crs_dep_time"].str.zfill(4).replace({'2400':'0000'}))

flights["arr_date_time"] = arr_time
flights["dep_date_time"] = dep_time

flights['arr_month'] = flights['arr_date_time'].dt.month

def reduce_size(data):
    return data.sample(frac=0.03,random_state=3245)
flights_sampled = flights.groupby(by=["op_unique_carrier","arr_month","dest_airport_id"]).apply(reduce_size).reset_index(drop=True)

flights_sampled.shape

flights_sampled = flights_sampled.drop(columns=["Unnamed: 0"])

flights_sampled.head()

flights_sampled.to_csv(r"../data/1_sampled data/3_percent.csv",index=False)
