==> notebooks/4_web_scraping.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support.expected_conditions import presence_of_element_located
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import NoSuchElementException


import pandas as pd
import numpy as np

https://www.airnav.com/airport/KXWA
https://adip.faa.gov/agis/public/#/airportSearch

flights_fe = pd.read_csv(r"../data/sample_with_inbound_data.csv")

airport_codes = flights_fe.origin.replace({'FCA':'GPI','ISN':'XWA','SCE':'UNV','BKG':'BBG','AZA':'IWA','USA':'JQF',
                      'MQT':'SAW','YUM':'NYL','SPN':'GSN','HHH':'HXD'}).unique()

['ADK', 'FCA', 'ISN', 'SCE', 'JHM', 'BKG', 'AZA', 'USA', 'MQT',
       'YUM', 'SPN', 'HHH']

['PADK','GPI','XWA','UNV','PHJH','BBG','IWA','JQF','SAW',
'NYL','GSN','HXD']


# airports_not_avilable_in_faa = ['FCA', 'ISN', 'SCE', 'BKG', 'AZA', 'USA', 'MQT', 'YUM', 'SPN', 'HHH']

# new_airport_codes = set(airport_codes).difference(airports_not_avilable_in_faa)

css_selectors = ["body > main:nth-child(2) > div:nth-child(1) > div:nth-child(1) > div:nth-child(3) > div:nth-child(1) > airport-details:nth-child(1) > div:nth-child(4) > div:nth-child(1) > div:nth-child(1) > div:nth-child(2) > div:nth-child(1) > div:nth-child(1) > div:nth-child(2)",
                 "div[ng-if='airportDetailData.npiasNumber'] div[class='col-md-8 ng-binding']",
                "div[ng-if='airportDetailData.serviceLevel'] div[class='col-md-8 ng-binding']",
                "div[ng-if='airportDetailData.hubType'] div[class='col-md-8 ng-binding']",
                "div[ng-if='airportDetailData.status'] div[class='col-md-8 ng-binding']",
                "span[ng-bind='ddToDms(airportDetailData.arpLatitude, true, false)']",
                 "span[ng-bind='ddToDms(airportDetailData.arpLongitude, true, true)']",
                "body > main:nth-child(2) > div:nth-child(1) > div:nth-child(1) > div:nth-child(3) > div:nth-child(1) > airport-details:nth-child(1) > div:nth-child(4) > div:nth-child(1) > div:nth-child(1) > div:nth-child(2) > div:nth-child(1) > div:nth-child(7) > div:nth-child(2)",
                "div[class='card mb-2 ng-scope'] div[class='row'] div:nth-child(1) div:nth-child(8) div:nth-child(2)",
                "p[class='m-0 ng-binding']",
                 "div[ng-if='airportDetailData.magneticVariation'] div[class='col-md-8 ng-binding']",
                 "div[class='card-body py-3'] div:nth-child(1) div:nth-child(11) div:nth-child(2)",
                "div[ng-if='airportDetailData.ownershipType'] div[class='col-md-8 ng-binding']",
                "div[ng-if='airportDetailData.facilityUse'] div[class='col-md-8 ng-binding']",
                 "div[ng-if='airportDetailData.areaInAcres'] div[class='col-md-8 ng-binding']",
                "div[ng-if='airportDetailData.tieInFssFlag'] div[class='col-md-8 ng-binding']",
                "div[ng-if='airportDetailData.beaconLensColor'] div[class='col-md-8 ng-binding']",
                "body > main:nth-child(2) > div:nth-child(1) > div:nth-child(1) > div:nth-child(3) > div:nth-child(1) > airport-details:nth-child(1) > div:nth-child(4) > div:nth-child(2) > div:nth-child(2) > div:nth-child(1) > div:nth-child(2) > div:nth-child(1) > div:nth-child(2)",
                "body > main:nth-child(2) > div:nth-child(1) > div:nth-child(1) > div:nth-child(3) > div:nth-child(1) > airport-details:nth-child(1) > div:nth-child(4) > div:nth-child(2) > div:nth-child(2) > div:nth-child(1) > div:nth-child(2) > div:nth-child(6) > div:nth-child(2)",
                "body > main:nth-child(2) > div:nth-child(1) > div:nth-child(1) > div:nth-child(3) > div:nth-child(1) > airport-details:nth-child(1) > div:nth-child(4) > div:nth-child(2) > div:nth-child(2) > div:nth-child(1) > div:nth-child(2) > div:nth-child(7) > div:nth-child(2)"]

variables = ["site_id","npias_no","service_level","hub_type","airport_status","lat","lon","city","county","elevation",
             "variation","last_inspection_date",
             "ownership_status","facility_use","area","fss_on_airport","beacon_color","air_carrier_annual",
            "total_operations_annual","operations_annual_ending","num_runways","avg_runway_length","avg_runway_width"]

#This example requires Selenium WebDriver 3.13 or newer
with webdriver.Chrome() as driver:
    new_airport_details = []
    for airport_code in airport_codes:
        url = f"https://adip.faa.gov/agis/public/#/airportData/{airport_code}"
        driver.implicitly_wait(10) # seconds
        driver.get(url)
        results = []
        for selector in css_selectors:
            try:
                result = driver.find_element_by_css_selector(selector)
                results.append(result.text)
            except NoSuchElementException as exc:
                print(f"{airport_code} : {exc}")
                result = ""
                results.append(result)
        we = driver.find_elements_by_css_selector("div[title='Dimensions, surfaces, edge lights, weight limits']")
        length=[]
        width = []
        num_runways = len(we)
        for runway in we:
            try:
                runway_dims = runway.find_element_by_xpath(".//div[contains(text(),'ft.')]").text
                dims = runway_dims.split("x")
                length.append(float(dims[0].rstrip(" ft. ")))
                width.append(float(dims[1].rstrip(" ft. ")))
            except NoSuchElementException:
                num_runways = num_runways-1
        results.append(num_runways)
        results.append(np.mean(length))
        results.append(np.mean(width))
        details = dict(zip(variables,results))
        details["airport_code"] = airport_code
        new_airport_details.append(details)

data = pd.DataFrame(new_airport_details)

data.head()

data.elevation = data.elevation.map(lambda x: x.rstrip("ft. "))
data.area = data.area.map(lambda x: x.rstrip("ac."))
data.air_carrier_annual = data.air_carrier_annual.map(lambda x: x.replace(',',''))
data.total_operations_annual = data.total_operations_annual.map(lambda x: x.replace(',',''))

data.head()

enplanements = pd.read_excel("../data/preliminary-cy20-commercial-service-enplanements.xlsx")

enplanements.head()

test_enplanements = enplanements[["Rank","CY 19 Enplanements","Locid"]]

test_enplanements.Locid = test_enplanements.Locid.replace({'FCA':'GPI','ISN':'XWA','SCE':'UNV','BKG':'BBG','AZA':'IWA','USA':'JQF',
                      'MQT':'SAW','YUM':'NYL','SPN':'GSN','HHH':'HXD'})

data = pd.merge(left=data,right=test_enplanements,how="left",left_on="airport_code",right_on="Locid")

data.to_csv(r"../data/faa_airport_details.csv",index=False)

//h3[normalize-space()='Runway Information']

with webdriver.Chrome() as driver:
    airport_code = "JFK"
    url = f"https://adip.faa.gov/agis/public/#/airportData/{airport_code}"
    driver.implicitly_wait(10) # seconds
    driver.get(url)


dims = "12079 ft. x 200 ft."
for dim in dims:
    print(dim.rstrip(" ft. "))
