==> notebooks/5_feature_engineering.ipynb <==
# Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from datetime import timedelta, datetime

a = pd.DataFrame({'A':[1,2,11],'B':[8,2,-8]})

(a>0).any(axis=1)

# Read Sample

# flights_sample = pd.read_csv(r"../data/1_sampled data/3_percent.csv")
flights_sample = pd.read_csv(r"flights_test.csv")
flights_sample.head()

flights_sample.info()

flights_sample.loc[:,["fl_date","crs_arr_time","crs_dep_time"]] = flights_sample[["fl_date","crs_arr_time","crs_dep_time"]].astype('string')

arr_time = pd.to_datetime(flights_sample["fl_date"]+" "+ flights_sample["crs_arr_time"].str.zfill(4).replace({'2400':'0000'}))
dep_time = pd.to_datetime(flights_sample["fl_date"]+" "+ flights_sample["crs_dep_time"].str.zfill(4).replace({'2400':'0000'}))

flights_sample["arr_date_time"] = arr_time
flights_sample["dep_date_time"] = dep_time

flights_sample = flights_sample.astype({'arr_date_time':'datetime64[ns]','dep_date_time':'datetime64[ns]'})

flights_sample.info()

flights_sample.shape

flights_sample = flights_sample[(flights_sample.arr_date_time>=datetime(2020,1,1))& (flights_sample.arr_date_time<=datetime(2020,1,7))]

flights_sample.shape

# Feature Engineering

flights_fe = flights_sample.copy()

## Airport Type

origin_airport = flights_fe["origin_airport_id"].value_counts()
dest_airport = flights_fe["dest_airport_id"].value_counts()

airports = pd.concat([origin_airport,dest_airport],axis=1)
airports = airports.sort_values(by=["origin_airport_id"],ascending=False)

df1, df2, df3, df4, df5 = np.array_split(airports, 5)

df1["Type"] = "Very Large"
df2["Type"] = "Large"
df3['Type'] = "Medium"
df4["Type"] = "Small"
df5["Type"] = "Very Small"

airport_type = pd.concat([df1,df2,df3,df4,df5]).drop(columns=["origin_airport_id","dest_airport_id"]).to_dict()['Type']

flights_fe["origin_airport_type"] = flights_fe.origin_airport_id.map(lambda x: airport_type[x] )
flights_fe["dest_airport_type"] = flights_fe.dest_airport_id.map(lambda x: airport_type[x] )

## number of inbound flights

def get_num_inbound_flights_carrier(data,minutes=-60,percentages_to_print=[],num_row_processed=[0]):

    if num_row_processed[0] in percentages_to_print:
        print(f"{datetime.now().time()} : {100*num_row_processed[0]/total_num}% rows processed")
    
    num_row_processed[0] = num_row_processed[0]+1
    carrier = data.mkt_unique_carrier
    airport = data.origin_airport_id
    date = data.dep_date_time
    fl_num = data.op_carrier_fl_num
    date_diff = date + timedelta(minutes=minutes)

    mask = ( 
    (flights_fe.mkt_unique_carrier==carrier) & (flights_fe.dest_airport_id==airport) & 
    (flights_fe.arr_date_time >= date_diff) & (flights_fe.arr_date_time <= date) & (flights_fe.op_carrier_fl_num != fl_num)
       )
    return flights_fe[mask].shape[0]


num_row_processed = [0]
total_num = flights_fe.shape[0]
percentages = np.round(np.linspace(0,total_num,11))
flights_fe["inbound_flights"] = flights_fe.apply(get_num_inbound_flights_carrier,axis=1,percentages_to_print=percentages,num_row_processed = num_row_processed)

flights_fe.shape

## Number of flights scheduled to take off from the airport -15 mins from dep_time

def get_num_dep_flights(data,minutes=-30,percentages_to_print=[],num_row_processed=[0]):
    
    if num_row_processed[0] in percentages_to_print:
        print(f"{datetime.now().time()} : {100*num_row_processed[0]/total_num}% rows processed")
    num_row_processed[0] = num_row_processed[0]+1
    
    airport = data.origin
    date = data.dep_date_time
    date_diff = date + timedelta(minutes=minutes)

    mask = ( (flights_fe.origin ==airport) & 
    (flights_fe.dep_date_time >= date_diff) & (flights_fe.dep_date_time <= date)
       )
    return flights_fe[mask].shape[0]-1


num_row_processed = [0]
total_num = flights_fe.shape[0]
percentages = np.round(np.linspace(0,total_num,11))
flights_fe.loc[:,"dep_flights_count"] = flights_fe.apply(get_num_dep_flights,axis=1,percentages_to_print=percentages,num_row_processed = num_row_processed)

flights_fe.shape

2+3

## Number of flights scheduled to arrive to airport +/-20 mins of scheduled arr_time

pd.Data

def get_num_arr_flights(data,minutes=30,percentages_to_print=[],num_row_processed=[0]):
    
    if num_row_processed[0] in percentages_to_print:
        print(f"{datetime.now().time()} : {100*num_row_processed[0]/total_num}% rows processed")
    num_row_processed[0] = num_row_processed[0]+1
    
    airport = data.dest
    date = data.arr_date_time + timedelta(minutes=minutes)
    date_diff = date - timedelta(minutes=minutes)

    mask = ( (flights_fe.dest ==airport) & 
    (flights_fe.arr_date_time >= date_diff) & (flights_fe.arr_date_time <= date)
       )
    return flights_fe[mask].shape[0]-1

num_row_processed = [0]
total_num = flights_fe.shape[0]
percentages = np.round(np.linspace(0,total_num,11))
flights_fe.loc[:,"arr_flights_count"] = flights_fe.apply(get_num_arr_flights,axis=1,percentages_to_print=percentages,num_row_processed = num_row_processed)

flights_fe.shape

## State names

flights_fe["origin_state"] = flights_fe.origin_city_name.map(lambda x: x.split(", ")[1])
flights_fe["dest_state"] = flights_fe.dest_city_name.map(lambda x: x.split(", ")[1])

flights_fe.origin_state.unique()

flights_fe.shape

## Weather

https://smoosavi.org/datasets/lstw

# # Drop Alska, hawaii, Puerto Rico, U.S. Virgin Islands and American Samoa = AK,HI, DC, GU, MH, MH, MP, PR, VI
# flights_fe = flights_fe[(~flights_fe.origin_state.isin(["AK","HI","DC","PR","VI","TT"])) &
#             (~flights_fe.dest_state.isin(["AK","HI","DC","PR","VI","TT"]))]

weather_kaggle = pd.read_csv(r"../data/external/WeatherEvents_Jan2016-Dec2020.csv")

weather_kaggle = weather_kaggle.astype({'StartTime(UTC)':'datetime64[ns]',
                                       'EndTime(UTC)':'datetime64[ns]'})

weather_kaggle = weather_kaggle[(weather_kaggle["StartTime(UTC)"]>datetime(year=2019,month=1,day=1))&
                                (weather_kaggle["EndTime(UTC)"]<datetime(year=2020,month=2,day=1))]

weather_kaggle.head()

flights_fe.origin = flights_fe.origin.replace({'FCA':'GPI','SCE':'UNV','BKG':'BBG','AZA':'IWA','USA':'JQF',
                      'MQT':'SAW','YUM':'NYL','HHH':'HXD'})
flights_fe.dest = flights_fe.dest.replace({'FCA':'GPI','SCE':'UNV','BKG':'BBG','AZA':'IWA','USA':'JQF',
                      'MQT':'SAW','YUM':'NYL','HHH':'HXD'})

def get_weather(data,percentages_to_print=[],num_row_processed=[0]):
    if num_row_processed[0] in percentages_to_print:
        print(f"{datetime.now().time()} : {100*num_row_processed[0]/total_num}% rows processed")
    num_row_processed[0] = num_row_processed[0]+1
    
    arr_time = data.arr_date_time
    dep_time = data.dep_date_time
    origin_id = data.origin
    dest_id = data.dest
    
    weather_details = pd.Series(dtype='string')
    
    try:
        weather_details["arr_type"] ,weather_details["arr_severity"] = weather_kaggle[(weather_kaggle["StartTime(UTC)"]<arr_time)&
                  (weather_kaggle["AirportCode"]=="K"+dest_id)].sort_values(by='StartTime(UTC)',ascending=False).iloc[0][['Type','Severity']]
    except IndexError as e:
        weather_details["arr_type"] ,weather_details["arr_severity"] = np.nan,np.nan
        
    try:
         weather_details["dep_type"] ,weather_details["dep_severity"] = weather_kaggle[(weather_kaggle["StartTime(UTC)"]<dep_time)&
                  (weather_kaggle["AirportCode"]=="K"+origin_id)].sort_values(by='StartTime(UTC)',ascending=False).iloc[0][['Type','Severity']]
    except IndexError as e:
        weather_details["dep_type"] ,weather_details["dep_severity"] = np.nan, np.nan
    
    print(num_row_processed)
    return weather_details

# num_row_processed = [0]
# total_num = flights_fe.shape[0]
# percentages = np.round(np.linspace(0,total_num,11))
# weather_details = flights_fe.apply(get_weather,axis=1,percentages_to_print=percentages,num_row_processed = num_row_processed)

# weather_details.head()

# flights_fe = pd.merge(left=flights_fe,right=weather_details,left_index=True,right_index=True)

## Traffic

## City Population

data = pd.read_csv(r"../data/external/census_data.csv")

population = data.groupby("STNAME").sum()
population.head()

codes = pd.read_csv(r"../data/external/state codes.csv")

population = pd.merge(left=population,right=codes,right_on='Name',left_index=True)

population = population.set_index('State or Region Code')

population.loc["DC","POPESTIMATE2019"] = 705749
population.loc["VI","POPESTIMATE2019"] = 104578
population.loc["PR","POPESTIMATE2019"] = 3654474
population.loc["TT","POPESTIMATE2019"] = 1394969

flights_fe["origin_population"] = flights_fe.origin_state.map(lambda x: population.loc[x]["POPESTIMATE2019"])

flights_fe["dest_population"] = flights_fe.origin_state.map(lambda x: population.loc[x]["POPESTIMATE2019"])

## Enplanements

enplanemntes = pd.read_excel("../data/external/preliminary-cy20-commercial-service-enplanements.xlsx")

enplanemntes = enplanemntes[["Locid","CY 19 Enplanements"]]
enplanemntes.head()

ser = flights_fe["origin"] 

flights_fe = pd.merge(left=flights_fe,right=enplanemntes,how="left",left_on="origin",right_on="Locid")

flights_fe.shape

# Airport Location details

## Airport details (number of gates, waiting area etc.,)

airport_details = pd.read_csv(r"../data/external/faa_airport_details.csv")
airport_details = airport_details.set_index("airport_code")
airport_details.area = airport_details.area.replace({"":np.nan})
airport_details.head()

airport_details = airport_details.drop(columns=['site_id', 'npias_no','airport_status','lat', 'lon', 'city', 'county',
                         'variation',
       'last_inspection_date','operations_annual_ending','elevation','air_carrier_annual','Locid','CY 19 Enplanements'])
airport_details.head()

flights_fe = pd.merge(left=flights_fe,right=airport_details,right_index=True,left_on='origin')
flights_fe = pd.merge(left=flights_fe,right=airport_details,right_index=True,left_on='dest',suffixes=("_or","_des"))

flights_fe.shape

## Date, hour, week

# flights_modified['arr_year'] = flights_modified['arr_date_time'].dt.year
flights_fe['arr_month'] = flights_fe['arr_date_time'].dt.month
flights_fe['arr_day_of_week'] = flights_fe['arr_date_time'].dt.day_of_week
flights_fe['arr_hour'] = flights_fe['arr_date_time'].dt.hour


# flights_modified['dep_year'] = flights_modified['dep_date_time'].dt.year
flights_fe['dep_month'] = flights_fe['dep_date_time'].dt.month
flights_fe['dep_day_of_week'] = flights_fe['dep_date_time'].dt.day_of_week
flights_fe['dep_hour'] = flights_fe['dep_date_time'].dt.hour

flights_fe.shape

## Carrier Profits in 2019

## Carrier employment satisfaction

## carrier cargo tie-ups

## Save the file

flights_fe.to_csv(r"../data/interim/2_feature_engineered/test_data.csv",index=False)
