==> notebooks/6.1_modelling_best_performance.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime,timedelta
import seaborn as sns
import matplotlib.pyplot as plt
# Import stats from scipy library
from scipy import stats
import statsmodels.api as sm


from sklearn.model_selection import train_test_split,GridSearchCV,RandomizedSearchCV
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler,MinMaxScaler
from sklearn.linear_model import LinearRegression

from sklearn.preprocessing import PolynomialFeatures

from sklearn.metrics import r2_score,mean_absolute_error,mean_squared_error
from sklearn.feature_selection import SequentialFeatureSelector

# Including weather data

flights_fe = pd.read_csv(r"../data/interim/2_feature_engineered/train_data_Akhmet.csv")

flights_fe.head()

df = flights_fe.copy()

X,y= df.drop(columns=["arr_delay"]),df[["arr_delay"]]

X_dummy = X.drop(columns=["op_carrier_fl_num",'beacon_color_or','beacon_color_des','ownership_status_des','facility_use_des','ownership_status_or','facility_use_or'])

X_dummy = pd.get_dummies(
    data=X_dummy,columns=["mkt_unique_carrier","op_unique_carrier",
                          'arr_month','arr_day_of_week','arr_hour','dep_month','dep_day_of_week','dep_hour', 
                          
                          'service_level_or','hub_type_or','fss_on_airport_or',
                          'service_level_des', 
                          'hub_type_des','fss_on_airport_des'  
                          
                          ,'origin_state','dest_state',
                          "weather_data"
                          ])

X_dummy.shape

X_train,X_test,y_train,y_test = train_test_split(X_dummy,y,train_size=0.7,random_state=0)

clf =LinearRegression()
clf.fit(X_train, y_train)
y_pred=clf.predict(X_test)

r2_score(y_test,y_pred)

mean_absolute_error(y_test,y_pred)

mean_squared_error(y_test,y_pred)

import xgboost as xgb
data_dmatrix = xgb.DMatrix(data=X_train,label=y_train)
xg_reg = xgb.XGBRegressor(objective ='reg:squarederror', colsample_bytree = 0.3, learning_rate = 0.1,
                max_depth = 5, alpha = 0.3, n_estimators = 1000)
xg_reg.fit(X_train,y_train)

y_pred = xg_reg.predict(X_test)
r2_score(y_test,y_pred)

## Excluding weather data

X,y= df.drop(columns=["arr_delay"]),df[["arr_delay"]]

X_dummy = X.drop(columns=["op_carrier_fl_num",'beacon_color_or','beacon_color_des','ownership_status_des','facility_use_des','ownership_status_or','facility_use_or',"weather_data",])

X_dummy = pd.get_dummies(
    data=X_dummy,columns=["mkt_unique_carrier","op_unique_carrier",
                          'arr_month','arr_day_of_week','arr_hour','dep_month','dep_day_of_week','dep_hour', 
                          
                          'service_level_or','hub_type_or','fss_on_airport_or',
                          'service_level_des', 
                          'hub_type_des','fss_on_airport_des'  
                          
                          ,'origin_state','dest_state'
                          ])

X_dummy.shape

X_train,X_test,y_train,y_test = train_test_split(X_dummy,y,train_size=0.7,random_state=0)

clf =LinearRegression()
clf.fit(X_train, y_train)
y_pred=clf.predict(X_test)

r2_score(y_test,y_pred)

mean_absolute_error(y_test,y_pred)

mean_squared_error(y_test,y_pred)

import xgboost as xgb
data_dmatrix = xgb.DMatrix(data=X_train,label=y_train)
xg_reg = xgb.XGBRegressor(objective ='reg:squarederror', colsample_bytree = 0.3, learning_rate = 0.1,
                max_depth = 5, alpha = 0.3, n_estimators = 1000)
xg_reg.fit(X_train,y_train)

y_pred = xg_reg.predict(X_test)
r2_score(y_test,y_pred)

X_test.shape

## Test data

test_data = pd.read_csv(r"../data/interim/2_feature_engineered/test_data.csv")

test = test_data.drop(columns=["op_carrier_fl_num",'beacon_color_or','beacon_color_des','ownership_status_des','facility_use_des','ownership_status_or','facility_use_or'])

drop_columns = list(set(test.columns)-set(X.columns))

X.shape

test.shape

test = test.drop(columns=drop_columns)

test.columns

test_dummy = pd.get_dummies(
    data=test,columns=["mkt_unique_carrier","op_unique_carrier",
                          'arr_month','arr_day_of_week','arr_hour','dep_month','dep_day_of_week','dep_hour', 
                          
                          'service_level_or','hub_type_or','fss_on_airport_or',
                          'service_level_des', 
                          'hub_type_des','fss_on_airport_des'  
                          
                          ,'origin_state','dest_state'
                          ])

add_columns = set(X_dummy.columns)-set(test_dummy.columns)
for col in add_columns:
    test_dummy[col] = 0

drop_columns =list(set(test_dummy.columns)-set(X_dummy.columns))

test_dummy = test_dummy.drop(columns=drop_columns)

y_pred = xg_reg.predict(test_dummy)

y_pred

results = test_data[["fl_date","mkt_carrier","mkt_carrier_fl_num","origin","dest"]]

results["predicted delay"] = y_pred

results.to_csv(r"../reports/predictions.csv")
