==> notebooks/7_modelling_classification.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime,timedelta
import seaborn as sns
from sklearn.preprocessing import StandardScaler

sns.heatmap(data.corr())

## Prepro

#df = pd.read_csv("train_data_Akhmet_v2.csv")
df = pd.read_csv("2_percent.csv")
print(df.shape)

df['cancelled'].value_counts()

#dfx = pd.read_csv("2_percent.csv")
#print(dfx.shape)

dfx['cancelled'].value_counts()

df=df[['mkt_unique_carrier', 'op_unique_carrier',
       'op_carrier_fl_num', 'arr_delay', 'crs_elapsed_time', 'distance',
       'inbound_flights', 'arr_month', 'arr_day_of_week',
       'arr_hour', 'dep_month', 'dep_day_of_week', 'dep_hour', 'hub_type_or',
       'area_or', 'total_operations_annual_or', 'hub_type_des',
       'facility_use_des', 'area_des', 'total_operations_annual_des',
       'origin_state', 'dest_state', 'cancelled', 'nas_delay',
       'security_delay', 'weather_delay', 'carrier_delay',
       'late_aircraft_delay','num_runways_des',
       'avg_runway_length_des', 'avg_runway_width_des','arr_type', 'arr_severity', 'dep_type', 'dep_severity']]
print(df.shape)
print(df.isnull().sum())

df.drop(columns=['nas_delay',
       'security_delay', 'weather_delay', 'carrier_delay',
       'late_aircraft_delay'],inplace=True)

#df=df.drop(columns=['arr_month', 'arr_day_of_week',
       #'arr_hour', 'dep_month', 'dep_day_of_week', 'dep_hour'])

#df.dropna(inplace=True)
df.fillna(0,inplace=True)
print(df.shape)

print(df.isnull().sum())

X,y= df.drop(columns=['cancelled',"arr_delay"]),df[['cancelled']]

X_dummy = X.drop(columns=["op_carrier_fl_num","crs_elapsed_time",])

X.columns

y['cancelled'].value_counts()
#ya= df[['nas_delay',
      # 'security_delay', 'weather_delay', 'carrier_delay',
       #'late_aircraft_delay']]
#yb= df['cancelled']
#ya=ya.fillna(0)

y['type']=y.apply(lambda x: x.idxmax(),axis=1)

y['type'].replace({"nas_delay": 0, "security_delay": 1,"weather_delay": 2,"carrier_delay": 3,"late_aircraft_delay": 4}, inplace=True)

#y=np.asarray(y['type'])
y=np.asarray(y['cancelled'])
print(y.shape)
print(X.shape)
print(X_dummy.shape)

X_dummy = pd.get_dummies(
    data=X_dummy,columns=["mkt_unique_carrier","op_unique_carrier",
                        "hub_type_or",
                          "hub_type_des",
                        "facility_use_des",
                        "origin_state","dest_state",'arr_type', 'arr_severity', 'dep_type', 'dep_severity',
                          ])
X_dummy=StandardScaler().fit_transform(X_dummy)
X_dummy=np.asarray(X_dummy)


#"arr_month","arr_day_of_week","arr_hour","dep_month","dep_hour","dep_day_of_week"

print(np.mean(y))
print(np.median(y))

print(y.shape)
print(X_dummy.shape)

from sklearn.decomposition import PCA
import plotly.express as px

X = X_dummy

pca = PCA(n_components=10)
principalComponents = pca.fit_transform(X)
print(principalComponents.shape)
print(type(principalComponents.shape))

#principalComponents=np.asarray(principalComponents)

## Modelling

from sklearn.model_selection import train_test_split,GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import r2_score,mean_absolute_error,mean_squared_error
from sklearn.preprocessing import PolynomialFeatures
from sklearn.pipeline import Pipeline
from imblearn.over_sampling import RandomOverSampler
from imblearn.over_sampling import SMOTE
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier


Multiclass Classification
The target variables are CARRIER_DELAY, WEATHER_DELAY, NAS_DELAY, SECURITY_DELAY, LATE_AIRCRAFT_DELAY. We need to do additional transformations because these variables are not binary but continuos. For each flight that was delayed, we need to have one of these variables as 1 and others 0.
It can happen that we have two types of delays with more than 0 minutes. In this case, take the bigger one as 1 and others as 0.

oversample = SMOTE(sampling_strategy='minority')
#oversample = SMOTE(sampling_strategy='all')
X1, y1 = oversample.fit_resample(X_dummy, y)

print(X1.shape)
print(y1.shape)

unique, counts = np.unique(y1, return_counts=True)
dict(zip(unique, counts))

X_train,X_test,y_train,y_test = train_test_split(X1,y1,train_size=0.7,random_state=0)

#clf = LogisticRegression(random_state=0)
#clf = LogisticRegression(random_state=0,max_iter=20000, multi_class='multinomial')

#clf = KNeighborsClassifier(n_neighbors=5)

#clf = GaussianNB()

#clf = RandomForestClassifier(max_depth=5, random_state=0)

#clf = SVC(kernel='poly',gamma='auto')

clf = LogisticRegression(random_state=0,max_iter=20000, multi_class='multinomial')
clf.fit(X_train, y_train)
y_pred=clf.predict(X_test)


print(type(y_test))
print(type(y_pred))

from sklearn.metrics import accuracy_score
from sklearn.metrics import recall_score
from sklearn.metrics import precision_score
from sklearn.metrics import f1_score
from sklearn.metrics import confusion_matrix
accuracy = accuracy_score(y_test,y_pred)
f1_score = f1_score(y_test,y_pred,average='weighted')
precision_score = precision_score(y_test,y_pred,average='weighted')
recall_score = recall_score(y_test,y_pred,average='micro')
cnf_matrix = confusion_matrix(y_test, y_pred)

#### Multi Logistic Regression SMOTE without Scaler
0.28330522765598654
0.25589846716567505
0.2746703349064258
0.28330522765598654
[[ 262  525  160  145  423]
 [ 255 1025   42   23  215]
 [ 225  525  211  248  342]
 [ 197  609  167  224  349]
 [ 182  573  149  171  462]]
 
#### Multi Logistic Regression SMOTE with Scaler
0.28330522765598654
0.25589846716567505
0.2746703349064258
0.28330522765598654
[[ 262  525  160  145  423]
 [ 255 1025   42   23  215]
 [ 225  525  211  248  342]
 [ 197  609  167  224  349]
 [ 182  573  149  171  462]]
 PCA MAKES THINGS WORST 10=33%Accuracy 100=45%


print(accuracy)
print(f1_score)
print(precision_score)
print(recall_score)
print(cnf_matrix)

import xgboost as xgb
data_dmatrix = xgb.DMatrix(data=X_train,label=y_train)
xg_reg = xgb.XGBRegressor(objective ='binary:hinge', colsample_bytree = 0.3, learning_rate = 0.1,
                max_depth = 2, alpha = 0.1, n_estimators = 1000)
xg_reg.fit(X_train,y_train)
y_pred = xg_reg.predict(X_test)

from sklearn.ensemble import RandomForestRegressor

from sklearn import metrics
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import GradientBoostingRegressor
param_grid={
    'alpha' :[0.1,0.3,0.5,0.7,1.0],
    'max_depth': [1,3,5,7,10],
    'n_estimators': [10,100,250,500,1000]
    }
params = {
    # Parameters that we are going to tune.
    'max_depth':6,
    'min_child_weight': 1,
    'eta':.3,
    'subsample': 1,
    'colsample_bytree': 1,
    # Other parameters
    'objective':'reg:squarederror',
}

## ModellingII
